Move `throughput` test from multipath.rs to tests/mod.rs
Also: Reduce test runtime by sending fewer
bytes in total, and make it more robust by
stepping forward until the stream datagram
arrives.
