+++ noq-proto/src/tests/multipath.rs
@@ -9,7 +9,6 @@ use assert_matches::assert_matches;
 use testresult::TestResult;
 use tracing::info;
 
-use crate::tests::util::{BwLimitConfig, BwLimitedRouting};
 use crate::{
     ClientConfig, ConnectionId, ConnectionIdGenerator, Endpoint, EndpointConfig, FourTuple,
     LOCAL_CID_COUNT, NetworkChangeHint, PathId, PathStatus, RandomConnectionIdGenerator,
@@ -2276,112 +2275,4 @@ fn regression_discarded_path_stats_are_up_to_date() -> TestResult {
     assert_ne!(discarded_stats.current_mtu, 0);
 
     Ok(())
-}
-
-#[test]
-fn throughput() -> TestResult {
-    const TOTAL_BYTES: usize = 10_000_000;
-    const BPS_LIMIT: u64 = 1_000_000;
-
-    let _guard = subscribe();
-    let mut pair = ConnPair::builder()
-        .with_routes(
-            BwLimitedRouting::new(
-                Pair::CLIENT_ADDR,
-                Pair::SERVER_ADDR,
-                crate::Instant::now(),
-                BwLimitConfig {
-                    bytes_per_second: BPS_LIMIT,
-                    buffer_size: 50 * 1500, // buffer that fits ~50 full packets
-                    latency: Duration::from_millis(3),
-                },
-            )
-            .into(),
-        )
-        .enable_multipath()
-        .connect();
-
-    let mut bytes_to_send = TOTAL_BYTES;
-    let mut bytes_received = 0;
-
-    let start = pair.time;
-    let client_stream = pair.conn_mut(Client).streams().open(Dir::Bi).unwrap();
-    // send the first batch to ensure the other side created the stream
-    bytes_to_send -= pair
-        .conn_mut(Client)
-        .send_stream(client_stream)
-        .write(&ZEROES)?;
-    pair.drive_client(); // generates the datagram
-    pair.advance_time(); // waits for latency for the datagram to be sent
-    pair.drive_server(); // receives the datagram
-    let server_stream = pair.conn_mut(Server).streams().accept(Dir::Bi).unwrap();
-    loop {
-        if bytes_to_send > 0 {
-            send_bytes(
-                pair.conn_mut(Client).send_stream(client_stream),
-                &mut bytes_to_send,
-            )?;
-            if bytes_to_send == 0 {
-                pair.conn_mut(Client).send_stream(client_stream).finish()?;
-            }
-        }
-        recv_bytes(
-            pair.conn_mut(Server).recv_stream(server_stream),
-            &mut bytes_received,
-        );
-        if !pair.step() {
-            break;
-        }
-    }
-
-    assert_eq!(bytes_to_send, 0);
-    assert_eq!(bytes_received, TOTAL_BYTES);
-
-    let time = pair.time.saturating_duration_since(start);
-    let bytes_per_second = TOTAL_BYTES as f64 / time.as_secs_f64();
-    info!(bytes_received, ?time, bytes_per_second);
-
-    let expected_bps = BPS_LIMIT as f64;
-    // Less than 2% deviation from the BPS limit
-    assert!(
-        (bytes_per_second - expected_bps).abs() / expected_bps < 0.05,
-        "deviated too far from expected throughput limit"
-    );
-
-    Ok(())
-}
-
-const ZEROES: [u8; 10_000] = [0u8; 10_000];
-
-fn send_bytes(mut send_stream: crate::SendStream<'_>, bytes_to_send: &mut usize) -> TestResult {
-    while *bytes_to_send > 10_000 {
-        match send_stream.write(&ZEROES) {
-            Ok(written) => {
-                *bytes_to_send -= written;
-            }
-            Err(crate::WriteError::Blocked) => return Ok(()),
-            Err(e) => panic!("{e:?}"),
-        }
-    }
-    while *bytes_to_send > 0 {
-        match send_stream.write(&vec![0u8; *bytes_to_send]) {
-            Ok(written) => {
-                *bytes_to_send -= written;
-            }
-            Err(crate::WriteError::Blocked) => return Ok(()),
-            Err(e) => panic!("{e:?}"),
-        }
-    }
-    Ok(())
-}
-
-fn recv_bytes(mut recv_stream: crate::RecvStream<'_>, bytes_received: &mut usize) {
-    let Ok(mut chunks) = recv_stream.read(true) else {
-        return;
-    };
-    while let Ok(Some(chunk)) = chunks.next(10_000) {
-        *bytes_received += chunk.bytes.len();
-    }
-    // The callee needs to immediately pair.step()
-    let _ = chunks.finalize();
 }
